Conflicts:
	code/modules/library/lib_machines.dm

+++ code/game/objects/items/weapons/janisigns.dm
@@ -12,13 +12,24 @@
 	origin_tech = "programming=3"
 	var/list/signs = list()
 	var/max_signs = 10
+	can_be_placed_into = list(
+		/obj/machinery/r_n_d/destructive_analyzer,
+		/obj/structure/table,
+		/obj/structure/rack,
+		/obj/structure/closet,
+		/obj/item/weapon/storage,
+		/obj/structure/safe,
+		/obj/machinery/disposal
+	)
 
 /obj/item/weapon/holosign_creator/afterattack(atom/target, mob/user, flag)
 	if(flag)
+		if(!check_allowed_items(target, 1)) return
 		var/turf/T = get_turf(target)
 		var/obj/effect/overlay/holograph/H = locate() in T
 		if(H)
 			user << "<span class='notice'>You use [src] to destroy [H].</span>"
+			signs.Remove(H)
 			qdel(H)
 		else
 			if(signs.len < max_signs)
